Changed ProcessingTimeInMS to be double instead of long

+++ src/CAAS/Models/Symmetric/Decryption/SymmetricDecryptionResponse.cs
@@ -12,6 +12,6 @@ namespace CAAS.Models.Symmetric.Decryption
 
         [Required]
         [DefaultValue(11)]
-        public long ProcessingTimeInMs { get; set; }
+        public double ProcessingTimeInMs { get; set; }
     }
 }
